refactor(Plugins-form): remove AppFrom, fix AppConfirm
