+++ src/core/components/Plugins/forms/Fields/AppFieldItems.vue
