+++ src/core/components/Plugins/forms/index.vue
@@ -69,8 +69,8 @@
 //var _Values = require('lodash/values');   // get values as array and object
 
 import Loader from './mixins/component-loading.js';
-import ChunkLoader from '../../../_mixins/loader-i18-chunk.js';
-import {helperValidationError} from '../../../Elements/AppErrorResponse/error-helper.js';
+import ChunkLoader from '../../_mixins/loader-i18-chunk.js';
+import {helperValidationError} from '../../Elements/AppErrorResponse/error-helper.js';
 
 import StoreDefault from './mixins/store.js';
 import Validator from './mixins/validator.js';
