+++ src/core/index.js
@@ -34,9 +34,10 @@ for (let NameComponent in CoreComponents) {
 }
 
 Vue.component('AppErrorResponse', ()=> import( /* webpackChunkName: "alerts" */ './components/Elements/AppErrorResponse/index.vue'));
+Vue.component('AppConfirm', ()=> import( /* webpackChunkName: "alerts" */ './components/Elements/AppConfirm.vue'));
 Vue.component('AppMessagesBlock', ()=> import( /* webpackChunkName: "alerts" */ './components/Elements/AppMessagesBlock'));
 
-Vue.component('SiteoPluginForms', ()=> import( /* webpackChunkName: "plugin-form" */ './components/Plugins/forms/AppForm/index.vue'));
+Vue.component('SiteoPluginForms', ()=> import( /* webpackChunkName: "plugin-form" */ './components/Plugins/forms/index.vue'));
 Vue.component('SiteoPluginSelectItems', ()=> import( /* webpackChunkName: "plugin-items" */ './components/Plugins/select-items/component.vue'));
